Test coreLibraryDesugaring for android6 support
