+++ sign-in-kotlin/app/src/androidTest/java/com/okta/sample/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.okta.oidc
+package com.okta.sample
 
 import androidx.test.InstrumentationRegistry
 import androidx.test.runner.AndroidJUnit4
